Fix slash commands (see comment)
Basically, because of how Discord has deprecated interaction response types 2 and 3, we cannot use the method we used previously. Some commands are still missing in this update but they will be added later. These commands are not necessary for the core features of the bot though, most of them are fun and can at least wait another day or two before getting added back.