quick fix for pr #14 (#15)
* Fix slash commands (see comment)

Basically, because of how Discord has deprecated interaction response types 2 and 3, we cannot use the method we used previously. Some commands are still missing in this update but they will be added later. These commands are not necessary for the core features of the bot though, most of them are fun and can at least wait another day or two before getting added back.

* Add permission table command

* Add "No VAD" role as a restriction

* Create handler for the about channel

* ESLint

* Make it so slash commands actually verify they're allowed to use the command 👀

* woops

This reverts commit 2ebded46dddfdc2e47b410f4c1c171a612a59451 except file src\handlers\slashCommands.js.

* Update about

* Allow the bot to actually see the channel when updating about

* feat: joinother slash-command

* fix: merge blocks together

* Quick fix

Co-authored-by: Stijn van der Kolk <stijn@kolkies.dev>
Co-authored-by: Project Blurple Access <blurple@v2202104146364150335.ultrasrv.de>